completed move from JSON to mongo-db
