+++ src/controllers/users.controllers.ts
@@ -8,7 +8,6 @@ export const getUsers = async (req: Request, res: Response) => {
         const users = await User.find({}, '_id username email role');
         res.status(200).json(users);
     } catch (error) {
-        console.error(error);
         res.status(500).send({ error: 'Error retrieving users' });
     }
 };
@@ -30,7 +29,6 @@ export const createUser = async (req: Request, res: Response) => {
         await user.save();
         res.status(201).json({ message: 'User created successfully', user});
     } catch (error) {
-        console.error(error);
         res.status(500).send({ error: 'Error creating user' });
     }
 };
