--- src/types/app.types.ts
@@ -1,9 +1,0 @@
-// Define for me a users type
-export type User = {
-    id: number;
-    username: string;
-    email: string;
-    role: string;
-    password: string;
-}
-// Define the Resources Type
